dedicated file for streammatcher
